Only try to upload stemcell when product needs one

+++ tasks/upload-product-and-stemcell/task.sh
@@ -4,23 +4,25 @@ if [[ ! -z "$NO_PROXY" ]]; then
   echo "$OM_IP $OPS_MGR_HOST" >> /etc/hosts
 fi
 
-FILE_PATH=`find ./pivnet-product -name *.pivotal`
-
 STEMCELL_VERSION=`cat ./pivnet-product/metadata.json | jq '.Dependencies[] | select(.Release.Product.Name | contains("Stemcells")) | .Release.Version'`
 
-echo "Downloading stemcell $STEMCELL_VERSION"
-pivnet-cli login --api-token="$PIVNET_API_TOKEN"
-pivnet-cli download-product-files -p stemcells -r $STEMCELL_VERSION -g $STEMCELL_GLOB --accept-eula
+if [ -n "$STEMCELL_VERSION" ]; then
+  echo "Downloading stemcell $STEMCELL_VERSION"
+  pivnet-cli login --api-token="$PIVNET_API_TOKEN"
+  pivnet-cli download-product-files -p stemcells -r $STEMCELL_VERSION -g $STEMCELL_GLOB --accept-eula
 
-SC_FILE_PATH=`find ./ -name *.tgz`
+  SC_FILE_PATH=`find ./ -name *.tgz`
 
-om-linux -t https://$OPS_MGR_HOST -u $OPS_MGR_USR -p $OPS_MGR_PWD -k upload-product -p $FILE_PATH
-
-om-linux -t https://$OPS_MGR_HOST -u $OPS_MGR_USR -p $OPS_MGR_PWD -k upload-stemcell -s $SC_FILE_PATH
-
-if [ ! -f "$SC_FILE_PATH" ]; then
+  if [ ! -f "$SC_FILE_PATH" ]; then
     echo "Stemcell file not found!"
-else
+    exit 1
+  fi
+
+  om-linux -t https://$OPS_MGR_HOST -u $OPS_MGR_USR -p $OPS_MGR_PWD -k upload-stemcell -s $SC_FILE_PATH
+
   echo "Removing downloaded stemcell $STEMCELL_VERSION"
   rm $SC_FILE_PATH
 fi
+
+FILE_PATH=`find ./pivnet-product -name *.pivotal`
+om-linux -t https://$OPS_MGR_HOST -u $OPS_MGR_USR -p $OPS_MGR_PWD -k upload-product -p $FILE_PATH
